api/requirements.txt
flask==2.3.2
gunicorn==20.1.0
google-cloud-bigquery==3.11.3